BUG Fixing images not being handled correctly when there's duplicates.
If a document contains duplicates, the existing code for handling
this doesn't work. One of the issues is the renaming of files if
one already exists. As it would turn out, this is not necessary
because LibreOffice names the files with a hash of the file
content anyway, so we'd never run into a case where you'd overwrite
an existing file as they're uniquely named.
e.g. "index-html-ab35fe.jpg".