FIX: Rename classes to something that makes sense
The logic for importing a document is wrapped by a load of FormFields that
concern themselves with the logic of performing various tasks to do with
the import more than they do with being FormFields. We'll rename these to
make a bit more sense with respect to the changes in API names and whatnot
since the days of SilverStripe 2.x where a lot of this seems to stem from.
Hopefully this will help folks (including the current maintainers) get
their heads around what goes where and does what.